archive/decode-rapl-v1/requirements.txt
# DECODE-RAPL Requirements
# Delayed Embedding and Coherent Dynamics for Emulating RAPL

# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
matplotlib>=3.7.0

# Optional: For generating architecture diagrams
torchviz>=0.0.2
graphviz>=0.20.0
